% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_countrycodes_function.R
\name{displaycountrylist}
\alias{displaycountrylist}
\title{World Bank Income Country codes}
\usage{
displaycountrylist()
}
\description{
This function allows you to get World Bank's country codes and names
}
\examples{
dataiwant <- displaycountrylist()
displaycountrylist()
}
\keyword{codes}
\keyword{country}
