% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_incomeclass_function.R
\name{wb_incomeclass}
\alias{wb_incomeclass}
\title{World Bank Income Classification function}
\usage{
wb_incomeclass(country = all, startyear = first, endyear = last)
}
\arguments{
\item{country}{Select countries by specifying country code. For example, to get classifications for India, type country = "IND". If you'd like multiple countries, type all country codes in c() function. Defaults to all. all is a vector of all country codes.}

\item{startyear}{To specify a range of years, you'll use startyear to specify the first year from which you want classifications. Defaults to first. first is a vector for the first value available for classifications.}

\item{endyear}{To specify a range of years, you'll use endyear to specify the latest year till which you want classifications. Defaults to last. last is a vector for the most recent value available for classifications.}
}
\description{
This function allows you to get World Bank's historical income classifications
}
\examples{
dataiwant <- wb_incomeclass(country = c("IND", "PAK", "BGD"), startyear = first, endyear = 2012)
wb_incomeclass()
}
\keyword{classification}
\keyword{income}
