% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_incomeclass_function.R
\name{wbincomeclass}
\alias{wbincomeclass}
\title{World Bank Income Classification function}
\usage{
wbincomeclass(country = "all", startyear = "first", endyear = "last")
}
\arguments{
\item{country}{Select countries by specifying country code. For example, to get classifications for India, type country = "IND". If you'd like multiple countries, type all country codes in c() function. Defaults to "all". "all" will give classification for all countries. To identify countrycodes, run the following code: d <- displaycountrylist().}

\item{startyear}{To specify a range of years, you'll use startyear to specify the first year from which you want classifications. Defaults to "first". "first" gives the first available year for classifications. The first year of available classifications is 1987.}

\item{endyear}{To specify a range of years, you'll use endyear to specify the latest year till which you want classifications. Defaults to "last". "last" gives the most recent year available for classifications. Before July 1, the most recent classification available will be 2 years prior to today. After July 1, the most recent classification available will be 1 year prior to today.}
}
\description{
This function allows you to get World Bank's historical income classifications
}
\examples{
dataiwant <- wbincomeclass(country = c("IND", "PAK", "BGD"), startyear = 1995, endyear = 2012)
dataimightwant <- wbincomeclass(country = "all", startyear = "first", endyear = "last")
dataiwillneed <- wbincomeclass(country = "FRA", startyear = "first", endyear = 2012)
wbincomeclass()
}
\keyword{classification}
\keyword{income}
