% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_incomeclass_function.R
\name{wbincomeclass}
\alias{wbincomeclass}
\title{World Bank Income Classification function}
\usage{
wbincomeclass(country = "all", startyear = 1987, endyear = 2019)
}
\arguments{
\item{country}{Select countries by specifying country code. For example, to get classifications for India, type country = "IND". If you'd like multiple countries, type all country codes in c() function. Defaults to all. all is a vector of all country codes. You can find official country codes here: https://unstats.un.org/unsd/tradekb/knowledgebase/country-code.}

\item{startyear}{To specify a range of years, you'll use startyear to specify the first year from which you want classifications. Defaults to first. first is a vector for the first value available for classifications. The first year of available classifications is 1987.}

\item{endyear}{To specify a range of years, you'll use endyear to specify the latest year till which you want classifications. Defaults to last. last is a vector for the most recent value available for classifications. Before July 1, the most recent classification available will be 2 years prior. After July, the most recent classification available will be 1 year prior.}
}
\description{
This function allows you to get World Bank's historical income classifications
}
\examples{
dataiwant <- wbincomeclass(country = c("IND", "PAK", "BGD"), startyear = 1995, endyear = 2012)
data <- wbincomeclass(country = all, startyear = first, endyear = last)
data <- wbincomeclass(country = "FRA", startyear = first, endyear = 2012)
wbincomeclass()
}
\keyword{classification}
\keyword{income}
